moved init code generation into module
