+++ src/genn/genn/code_generator/customUpdateGroupMerged.cc
@@ -5,7 +5,6 @@
 
 using namespace CodeGenerator;
 
-
 //--------------------------------------------------------------------------
 // Anonymous namespace
 //--------------------------------------------------------------------------
