+++ src/genn/genn/genn.vcxproj
@@ -25,13 +25,13 @@
     <ClCompile Include="code_generator\codeStream.cc" />
     <ClCompile Include="code_generator\customUpdateGroupMerged.cc" />
     <ClCompile Include="code_generator\generateAll.cc" />
-    <ClCompile Include="code_generator\generateInit.cc" />
     <ClCompile Include="code_generator\generateMakefile.cc" />
     <ClCompile Include="code_generator\generateMSBuild.cc" />
     <ClCompile Include="code_generator\generateRunner.cc" />
     <ClCompile Include="code_generator\generateSupportCode.cc" />
     <ClCompile Include="code_generator\generateSynapseUpdate.cc" />
     <ClCompile Include="code_generator\groupMerged.cc" />
+    <ClCompile Include="code_generator\initGroupMerged.cc" />
     <ClCompile Include="code_generator\neuronUpdateGroupMerged.cc" />
     <ClCompile Include="code_generator\modelSpecMerged.cc" />
     <ClCompile Include="code_generator\presynapticUpdateStrategySIMT.cc" />
@@ -65,13 +65,13 @@
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\codeStream.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\customUpdateGroupMerged.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateAll.h" />
-    <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateInit.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateMakefile.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateMSBuild.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateRunner.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateSupportCode.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateSynapseUpdate.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\groupMerged.h" />
+    <ClInclude Include="..\..\..\include\genn\genn\code_generator\initGroupMerged.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\neuronUpdateGroupMerged.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\modelSpecMerged.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\presynapticUpdateStrategySIMT.h" />
